refactored estimator code into tf.keras code and normal estimator code
